Remove enum arithmetic
This becomes a warning in C++20, at least with gcc

+++ src/quadrature/quadrature_grid_1D.C
@@ -29,7 +29,7 @@ void QGrid::init_1D()
 {
   _points.resize(_order + 1);
   _weights.resize(_order + 1);
-  const Real dx = Real(2)/(_order+1);
+  const Real dx = Real(2)/(static_cast<int>(_order)+1);
   for (int i = 0; i != _order + 1; ++i)
     {
       _points[i](0) = (i+0.5)*dx-1;
